refactor: rename workspace folder to vault
- Rename packages/plugin-kit/src/workspace/ -> src/vault/
- Rename apps/remote-server/src/core/workspace/ -> src/core/vault/
- Update package.json exports ./workspace -> ./vault
- Update tsup entry workspace -> vault
- Update all import paths to use plugin-kit/vault and ./vault/
