+++ apps/remote-server/src/core/engine-singleton.ts
@@ -1,7 +1,7 @@
 import { Engine } from 'pulse-coder-engine';
 import { memoryIntegration } from './memory-integration.js';
 import { worktreeIntegration } from './worktree/integration.js';
-import { vaultIntegration } from './workspace/integration.js';
+import { vaultIntegration } from './vault/integration.js';
 import { cronJobTool } from './tools/cron-job.js';
 import { deferDemoTool } from './tools/defer-demo.js';
 import { jinaAiReadTool } from './tools/jina-ai.js';
