+++ apps/remote-server/src/index.ts
@@ -5,7 +5,7 @@ import { engine } from './core/engine-singleton.js';
 import { sessionStore } from './core/session-store.js';
 import { memoryIntegration } from './core/memory-integration.js';
 import { worktreeIntegration } from './core/worktree/integration.js';
-import { vaultIntegration } from './core/workspace/integration.js';
+import { vaultIntegration } from './core/vault/integration.js';
 import { devtoolsStore } from './core/devtools.js';
 import { startDiscordGateway } from './adapters/discord/gateway-manager.js';
 import { registerDiscordApplicationCommands } from './adapters/discord/application-commands.js';
