+++ packages/plugin-kit/src/index.ts
@@ -1,3 +1,3 @@
 export * from './worktree/index.js';
-export * from './workspace/index.js';
+export * from './vault/index.js';
 export * from './devtools/index.js';
