+++ packages/plugin-kit/src/vault/index.ts
