+++ packages/plugin-kit/src/vault/integration.ts
