+++ packages/plugin-kit/src/vault/service.ts
