+++ packages/plugin-kit/src/vault/tools.ts
