+++ packages/plugin-kit/src/vault/types.ts
