+++ packages/plugin-kit/tsup.config.ts
@@ -4,7 +4,7 @@ export default defineConfig({
   entry: {
     index: 'src/index.ts',
     worktree: 'src/worktree/index.ts',
-    workspace: 'src/workspace/index.ts',
+    vault: 'src/vault/index.ts',
     devtools: 'src/devtools/index.ts',
   },
   format: ['esm', 'cjs'],
